Add generator, transformer, and validators for better configurability
