+++ asgi_correlation_id/middleware.py
@@ -1,30 +1,33 @@
 import logging
-from dataclasses import dataclass
-from uuid import UUID, uuid4
+from dataclasses import dataclass, field
+from typing import Callable, List
+from uuid import uuid4
 
 from starlette.datastructures import Headers
 from starlette.types import ASGIApp, Message, Receive, Scope, Send
 
 from asgi_correlation_id.context import correlation_id
 from asgi_correlation_id.extensions.sentry import get_sentry_extension
+from asgi_correlation_id.validators import is_valid_uuid
 
 logger = logging.getLogger('asgi_correlation_id')
-
-
-def is_valid_uuid(uuid_: str) -> bool:
-    """
-    Check whether a string is a valid v4 uuid.
-    """
-    try:
-        return bool(UUID(uuid_, version=4))
-    except ValueError:
-        return False
 
 
 @dataclass
 class CorrelationIdMiddleware:
     app: ASGIApp
     header_name: str = 'X-Request-ID'
+
+    # ID generating function
+    generator: Callable[[], str] = field(default=lambda: uuid4().hex)
+
+    # Validators for discarding badly formatted IDs
+    validators: List[Callable[[str], bool]] = None  # type: ignore[assignment]
+
+    # Extra handler layer for mutating IDs if needed
+    transformer: Callable[[str], str] = field(default=lambda a: a)
+
+    # Deprecated and will be removed in v2, use validators instead
     validate_header_as_uuid: bool = True
 
     async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
@@ -35,15 +38,15 @@ class CorrelationIdMiddleware:
             await self.app(scope, receive, send)
             return
 
-        header_value = Headers(scope=scope).get(self.header_name.lower())
+        header_value: str = Headers(scope=scope).get(self.header_name.lower())
 
         if not header_value:
-            id_value = uuid4().hex
-        elif self.validate_header_as_uuid and not is_valid_uuid(header_value):
-            logger.warning('Generating new UUID, since header value \'%s\' is invalid', header_value)
-            id_value = uuid4().hex
+            id_value: str = self.transformer(self.generator())
+        elif self.validators and not all(validator(header_value) for validator in self.validators):
+            logger.warning("Generating new ID, since header value '%s' is invalid", header_value)
+            id_value = self.transformer(self.generator())
         else:
-            id_value = header_value
+            id_value = self.transformer(header_value)
 
         correlation_id.set(id_value)
         self.sentry_extension(id_value)
@@ -67,6 +70,9 @@ class CorrelationIdMiddleware:
         If Sentry is installed, propagate correlation IDs to Sentry events.
         If Celery is installed, propagate correlation IDs to spawned worker processes.
         """
+        if self.validators is None:
+            self.validators = [is_valid_uuid] if self.validate_header_as_uuid else []
+
         self.sentry_extension = get_sentry_extension()
         try:
             import celery  # noqa: F401
