+++ asgi_correlation_id/validators.py
@@ -1,0 +1,11 @@
+from uuid import UUID
+
+
+def is_valid_uuid(uuid_: str) -> bool:
+    """
+    Check whether a string is a valid v4 uuid.
+    """
+    try:
+        return bool(UUID(uuid_, version=4))
+    except ValueError:
+        return False
